requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
websockets>=12.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Football Data Scraping
soccerdata>=1.7.0
ScraperFC>=2.0.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
feedparser>=6.0.10
httpx>=0.26.0

# LLM Integration
ollama>=0.1.7

# Configuration
pyyaml>=6.0.1
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Templating
jinja2>=3.1.3

# Utilities
python-dateutil>=2.8.2
aiofiles>=23.2.1
setuptools>=69.0.0  # Provides distutils for Python 3.12+

# Background Tasks
apscheduler>=3.10.4
